Server and Client in extra classes
