+++ app/src/main/java/com/example/scotland_yard_board_game/Client.java
@@ -2,15 +2,19 @@ package com.example.scotland_yard_board_game;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.os.AsyncTask;
 import android.os.Bundle;
+import android.util.Log;
 
 
-public class Client extends AppCompatActivity {
+
+public class Client extends AsyncTask<Void, Void, Void> {
+    private static final String TAG = "Client";
+
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_client);
+    protected Void doInBackground(Void... voids) {
+        Log.d(TAG, "doInBackground: Call the Server");
 
+        return null;
     }
-    
 }
